disable all handlers except close
