+++ program/Cargo.toml
@@ -14,7 +14,6 @@ bincode = "1.3.3"
 borsh = "0.9.3"
 num-derive = "0.3.3"
 num-traits = "~0.2"
-mpl-token-metadata = { version = "1.8.0", features = ["no-entrypoint"] }
 mpl-utils = "0.0.5"
 serde = { version = "1.0", optional = true }
 serde_with = { version = "1.14.0", optional = true }
@@ -47,5 +46,3 @@ features = ["precommit-hook", "user-hooks"]
 
 [profile.release]
 strip = true  # Automatically strip symbols from the binary.
-codegen-units = 1  # Optimize for small binary size.
-lto = true  # Use link-time optimization.
