+++ program/src/lib.rs
@@ -10,5 +10,4 @@ use solana_program::{pubkey, pubkey::Pubkey};
 
 solana_program::declare_id!("migrxZFChTqicHpNa1CAjPcF29Mui2JU2q4Ym7qQUTi");
 
-const MPL_TOKEN_AUTH_RULES_ID: Pubkey = pubkey!("auth9SigNpDKz4sJJ1DfCTuZrZNSAgh9sFD3rboVmgg");
 pub const PROGRAM_SIGNER: Pubkey = pubkey!("4fDQAj27ahBfXw3ZQumg5gJrMRUCzPUW6RxrRPFMC8Av");
