+++ program/src/processor/mod.rs
@@ -1,43 +1,19 @@
-use crate::{
-    errors::MigrationError,
-    instruction::{InitializeArgs, MigrationInstruction, UpdateArgs},
-    state::{MigrationState, ProgramSigner, UnlockMethod, MIGRATION_WAIT_PERIOD, SPL_TOKEN_ID},
-};
-use borsh::{BorshDeserialize, BorshSerialize};
-use mpl_token_metadata::{
-    instruction::{builders::MigrateBuilder, InstructionBuilder, MigrateArgs},
-    state::{Metadata, MigrationType, TokenMetadataAccount, EDITION, PREFIX},
-};
+use crate::{errors::MigrationError, instruction::MigrationInstruction, state::MigrationState};
+use borsh::BorshDeserialize;
 use mpl_utils::{assert_derivation, assert_owned_by, assert_signer};
 use solana_program::{
     account_info::{next_account_info, AccountInfo},
-    clock::Clock,
     entrypoint::ProgramResult,
-    program::invoke_signed,
     program_error::ProgramError,
-    program_memory::sol_memcpy,
-    program_option::COption,
     pubkey::Pubkey,
-    sysvar::{self, Sysvar},
 };
 
-use spl_token::state::Account as TokenAccount;
-
 mod close;
-mod initialize;
-mod migrate;
 mod misc;
-mod start;
 mod update;
-mod validators;
+mod validator;
 
 use close::close_migration_state;
-use initialize::initialize_migration;
-use migrate::migrate_item;
-use misc::init_signer;
-use start::start_migration;
-use update::update_state;
-use validators::*;
 
 pub struct Processor;
 impl Processor {
@@ -50,14 +26,14 @@ impl Processor {
             MigrationInstruction::try_from_slice(instruction_data)?;
 
         match instruction {
-            MigrationInstruction::Initialize(args) => {
-                initialize_migration(program_id, accounts, args)
+            MigrationInstruction::Initialize(_) => {
+                Err(MigrationError::MigrationWindowClosed.into())
             }
-            MigrationInstruction::Update(args) => update_state(program_id, accounts, args),
+            MigrationInstruction::Update(_) => Err(MigrationError::MigrationWindowClosed.into()),
             MigrationInstruction::Close => close_migration_state(program_id, accounts),
-            MigrationInstruction::Start => start_migration(program_id, accounts),
-            MigrationInstruction::Migrate => migrate_item(program_id, accounts),
-            MigrationInstruction::InitSigner => init_signer(program_id, accounts),
+            MigrationInstruction::Start => Err(MigrationError::MigrationWindowClosed.into()),
+            MigrationInstruction::Migrate => Err(MigrationError::MigrationWindowClosed.into()),
+            MigrationInstruction::InitSigner => Err(MigrationError::MigrationWindowClosed.into()),
         }
     }
 }
