+++ program/src/utils.rs
@@ -1,57 +1,6 @@
-use mpl_token_metadata::{
-    assertions::collection::assert_is_collection_delegated_authority,
-    state::{CollectionAuthorityRecord, Metadata},
-};
 use solana_program::{
-    account_info::AccountInfo, entrypoint::ProgramResult, program_error::ProgramError,
-    pubkey::Pubkey, system_program,
+    account_info::AccountInfo, entrypoint::ProgramResult, pubkey::Pubkey, system_program,
 };
-
-use crate::{errors::MigrationError, state::MigrationState, PROGRAM_SIGNER};
-
-pub fn assert_valid_delegate(
-    delegate_pubkey: &Pubkey,
-    delegate_record_info: &AccountInfo,
-    collection_metadata: &Metadata,
-    migration_state: &MigrationState,
-) -> Result<(), ProgramError> {
-    let info = &migration_state.collection_info;
-    // Mint is the correct one for the metadata account.
-    if collection_metadata.mint != info.mint {
-        return Err(MigrationError::MetadataMintMistmatch.into());
-    }
-
-    if collection_metadata.update_authority != info.authority {
-        return Err(MigrationError::InvalidAuthority.into());
-    }
-
-    let bump = assert_is_collection_delegated_authority(
-        delegate_record_info,
-        delegate_pubkey,
-        &info.mint,
-    )?;
-
-    let data = delegate_record_info.try_borrow_data()?;
-    if data.len() == 0 {
-        return Err(MigrationError::InvalidDelegate.into());
-    }
-
-    let record = CollectionAuthorityRecord::from_bytes(&data)?;
-
-    if record.bump != bump {
-        return Err(MigrationError::InvalidDelegate.into());
-    }
-
-    if let Some(update_authority) = record.update_authority {
-        if update_authority != collection_metadata.update_authority {
-            return Err(MigrationError::InvalidDelegate.into());
-        }
-    } else {
-        return Err(MigrationError::InvalidDelegate.into());
-    }
-
-    Ok(())
-}
 
 pub fn close_program_account<'a>(
     account_info: &AccountInfo<'a>,
@@ -75,15 +24,4 @@ pub fn close_program_account<'a>(
 pub fn find_migration_state_pda(mint: &Pubkey) -> (Pubkey, u8) {
     let seeds = &[b"migration", mint.as_ref()];
     Pubkey::find_program_address(seeds, &crate::ID)
-}
-
-pub fn find_delegate_record_pda(mint: &Pubkey) -> (Pubkey, u8) {
-    let seeds = &[
-        mpl_token_metadata::state::PREFIX.as_bytes(),
-        mpl_token_metadata::ID.as_ref(),
-        mint.as_ref(),
-        mpl_token_metadata::pda::COLLECTION_AUTHORITY.as_bytes(),
-        PROGRAM_SIGNER.as_ref(),
-    ];
-    Pubkey::find_program_address(seeds, &mpl_token_metadata::ID)
 }
